refactor: 마이페이지, 작가페이지 최적화 완료
