+++ src/components/exhibition-register/ExhibitionHeader.tsx
@@ -1,3 +1,4 @@
+import { ChevronLeft } from 'lucide-react';
 import { useNavigate } from 'react-router-dom';
 
 export function ExhibitionHeader() {
@@ -11,17 +12,9 @@ export function ExhibitionHeader() {
         className="cursor-pointer absolute left-5"
         aria-label="뒤로가기"
       >
-        <svg width="24" height="24" viewBox="0 0 24 24" fill="none">
-          <path
-            d="M15 5l-7 7 7 7"
-            stroke="#0f172a"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
+        <ChevronLeft />
       </button>
-      <h1 className="text-neutral-900 text-xl font-bold">전시 기본 정보</h1>
+      <div className="text-neutral-900 text-xl font-bold font-['Pretendard']">전시 기본 정보</div>
     </div>
   );
 }
