+++ src/components/mypage/index.ts
@@ -1,0 +1,6 @@
+export { ArtistCard } from './ArtistCard';
+export { ArtworkCard } from './ArtworkCard';
+export { AuthPageHeader } from './AuthPageHeader';
+export { ExhibitionCard } from './ExhibitionCard';
+export { MyPageHeader } from './MyPageHeader';
+export { SettingsSheet } from './SettingsSheet';
