+++ src/constants/index.ts
@@ -1,1 +1,3 @@
+export * from './authPage';
 export * from './exhibition';
+export * from './mypage';
