+++ src/utils/mypage.ts
@@ -1,0 +1,3 @@
+export function statusBadgeClass(status: string): string {
+  return status.includes('종료') ? 'bg-neutral-400' : 'bg-sky-600';
+}
